+++ R/nested_ensemble_average.R
@@ -11,7 +11,7 @@
 #'  nested modeltime table results
 #' @param model_ids A vector of id's (`.model_id`) identifying which submodels to
 #'  use in the ensemble.
-#' @param control Controls various aspects of the ensembling process. See [control_nested_fit()].
+#' @param control Controls various aspects of the ensembling process. See `modeltime::control_nested_fit()`.
 #'
 #' @returns
 #' The nested modeltime table with an ensemble model added.
@@ -336,7 +336,7 @@ ensemble_nested_average_parallel <- function(object,
         utils::capture.output() %>%
         stringr::str_remove("Time difference of ")
 
-    if (control$verbose) cli::cli_inform(stringr::str_glue("Finished in: {time_elapsed}."))
+    if (control$verbose) cli::cli_inform("Finished in: {time_elapsed}.")
 
     # STRUCTURE ----